portal-api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
azure-cosmos>=4.7.0
azure-identity>=1.17.0
python-jose[cryptography]>=3.3.0
structlog>=24.0.0
python-dotenv>=1.0.0
sse-starlette>=2.1.0
